refactor: DiaryContentを切り出し
